+++ moodBookProject/app/src/main/java/com/example/moodbook/ui/myFriendMoodMap/MyFriendMoodMapFragment.java
@@ -1,21 +1,39 @@
 package com.example.moodbook.ui.myFriendMoodMap;
 
+import android.app.Dialog;
+import android.graphics.Bitmap;
+import android.graphics.drawable.BitmapDrawable;
+import android.location.Location;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
+import com.example.moodbook.DBMoodSetter;
 import com.example.moodbook.DBUpdate;
+import com.example.moodbook.MoodMapFragment;
 import com.example.moodbook.Mood;
+import com.example.moodbook.MoodMapFragment;
 import com.example.moodbook.PageFragment;
 import com.example.moodbook.R;
+import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.MapView;
 import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.model.BitmapDescriptor;
+import com.google.android.gms.maps.model.BitmapDescriptorFactory;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
 
@@ -30,7 +48,7 @@ import java.util.ArrayList;
  *  This activity is used to view a where a users friends' moods take place on a map
  * @see DBUpdate
  */
-public class MyFriendMoodMapFragment extends PageFragment implements OnMapReadyCallback, DBUpdate {
+public class MyFriendMoodMapFragment extends MoodMapFragment implements OnMapReadyCallback {
     //private MyFriendMoodMapViewModel MyFriendMoodMapViewModel;
 
     ///// Member Variables /////
@@ -40,9 +58,11 @@ public class MyFriendMoodMapFragment extends PageFragment implements OnMapReadyC
     private FirebaseFirestore db;
     private FirebaseAuth mAuth;
     private String userID;
+    private DBMoodSetter dbMoodSetter;
 
+    @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View root = super.onCreateView(inflater, container, savedInstanceState, R.layout.fragment_friendmoodmap);
+        View root = inflater.inflate( R.layout.fragment_friendmoodmap, container, false);
 
         mapView = root.findViewById(R.id.friendMapView);
         mapView.onCreate(savedInstanceState);
@@ -59,6 +79,8 @@ public class MyFriendMoodMapFragment extends PageFragment implements OnMapReadyC
         mAuth = FirebaseAuth.getInstance();
         userID = mAuth.getUid();
 
+        dbMoodSetter = new DBMoodSetter(mAuth, getContext());
+
         return root;
     }
 
@@ -69,6 +91,25 @@ public class MyFriendMoodMapFragment extends PageFragment implements OnMapReadyC
 
         // update list of markers
         updateList(db);
+
+        mapView.setContentDescription("MAP READY");
+
+        moodMap.setOnMarkerClickListener(new GoogleMap.OnMarkerClickListener() {
+            @Override
+            public boolean onMarkerClick(Marker marker) {
+                // get mood from data list with by using tag
+                int pos = (int)marker.getTag();
+                Mood mood = moodDataList.get(pos);
+
+                // create dialog popup
+                Dialog dialog = new Dialog(getContext());
+
+                // bind mood data to dialog layout
+                bindViews(mood, dialog, dbMoodSetter).show();
+
+                return false;
+            }
+        });
     }
 
     @Override
@@ -96,17 +137,26 @@ public class MyFriendMoodMapFragment extends PageFragment implements OnMapReadyC
     }
 
     @Override
-    public void updateList(FirebaseFirestore db) {
-        // todo: implement this
+    public void updateList(final FirebaseFirestore db) {
         try {
             db.collection("USERS")
                     .document(userID)
                     .collection("FRIENDS")
                     .get()
-                    .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
+                    .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                         @Override
-                        public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
+                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                            if(task.isSuccessful()){
+                                for(int i = 0; i < task.getResult().size(); i++){
+                                    DocumentSnapshot doc = task.getResult().getDocuments().get(i);
 
+                                    if (doc.exists() && !(doc.getId().equals("null"))){
+                                        Log.i("DOCID", "id: " + doc.getId() + "  |  " + doc.get("uid"));
+                                        getFriendMood(db, (String) doc.get("uid"));
+
+                                    }
+                                }
+                            }
                         }
                     });
         } catch (Exception e){
@@ -115,8 +165,61 @@ public class MyFriendMoodMapFragment extends PageFragment implements OnMapReadyC
     }
 
     private void drawMoodMarkers(ArrayList<Mood> moodDataList){
-        // todo: implement this
+        int emotionResource;
+        LatLng moodLatLng;
+
+        int i;
+        for(i = 0; i < moodDataList.size(); i++){
+            Mood mood = moodDataList.get(i);
+            // get image resource for the mood marker
+            emotionResource = mood.getEmotionImageResource();
+
+            // get location of mood
+            Location moodLocation = mood.getLocation();
+            moodLatLng = new LatLng(moodLocation.getLatitude(), moodLocation.getLongitude());
+
+            // use png image resource as marker icon
+            BitmapDrawable bitmapDrawable = (BitmapDrawable) getResources().getDrawable(emotionResource);
+            Bitmap b = bitmapDrawable.getBitmap();
+            Bitmap smallMarker = Bitmap.createScaledBitmap(b, 100, 100, false);
+            BitmapDescriptor bitmapDescriptor = BitmapDescriptorFactory.fromBitmap(smallMarker);
+
+            // draw on map
+            moodMap.addMarker(new MarkerOptions().position(moodLatLng).icon(bitmapDescriptor)).setTag(i);
+
+            // zoom in and focus on the most recent mood, ie. the last mood in list
+            if(i+1 == moodDataList.size()){
+                moodMap.animateCamera(CameraUpdateFactory.newLatLngZoom(moodLatLng, 11.0f));
+            }
+
+        }
     }
 
+    private void getFriendMood(FirebaseFirestore db, String id) {
+        try {
+
+
+        db.collection("USERS")
+                .document(id)
+                .collection("MOODS")
+                .get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()) {
+                            DocumentSnapshot doc = task.getResult().getDocuments().get(task.getResult().size() - 1);
+                            if (doc.exists() && !(doc.getId().equals("null"))) {
+                                Log.i("mood ID", "id: " + doc.getId());
+
+                            }
+                        }
+                    }
+                });
+
+        } catch (Exception e){
+            e.printStackTrace();
+        }
+
+    }
 
 }
